created groupChat component to contian groupChatMessages and groupChatInput
